requirements.txt
discord.py
g4f
firebase-admin
python-dotenv
aiohttp
colorama